fix: take the libraries and unit_tests green on Linux GCC and Clang
The first delivery level of docs/linux-macos-port-assessment-2026-09-07.md:
basic, SpriteLib, TextSystem, dxlib, framelib, packetwire, gamemodel and
VS_UI compile and link on Linux, and unit_tests runs the same 601 tests
the Windows trees run - under GCC 13, Clang 18, and GCC with ASan and
UBSan, where UBSan now reports nothing. Two of the fixes are real
defects the Linux build found; the rest are one-line portability.

The defects. ShopVersion_t was `typedef long`, so on LP64 szShopVersion
was 8 and a Linux client sized GCShopBought 288 bytes against the
server's 284 - tests/wire-layout.txt caught it on the first run. It is
int32_t, as the server spells it; the wire was always four bytes. And
every palette sprite placed its scanline pointer table at
m_pData + m_Size, whatever byte the pixel data ended on, so every read
of the table was a misaligned pointer load - the 15 UBSan reports, in
CSpritePal, CAlphaSpritePal, CSpritePalBase and two tests. One helper,
CSpritePalBase::AllocateDataAndScanlineTable, pads the table up to the
pointer alignment, and the four loaders, the copy operator and the test
fixture go through it; test_cspritepalbase.cpp pins the alignment with
13 bytes of pixel data.

A link gap the key-function rule exposed: GCC emits MSerum's vtable in
gamemodel (its key function, GetMaxNumber, is in MItem.cpp) and that
vtable names the inherited MUsePotionItem::UseInventory, defined in the
executable's MItemUse.cpp, so unit_tests had an undefined reference.
MSVC emits vtables where an object is constructed and never saw it. The
member is defined in MItem.cpp now and reaches its body through a new
MItemHost slot, UsePotionFromInventory, which GameInit.cpp installs; a
test binary installs none and the use is a no-op. This is the seam
pattern the host already uses for the animation clock and the magazine.

Compile fixes: `enum Race Race[3]` in UserInformation.h and the same
member in VS_UI_title.h (GCC's -Wchanges-meaning is an error);
ItemTypes.h's `#include "RaceType.h"` becomes "../RaceType.h" - the bare
name resolved only through MSVC's include-stack search; `<VECTOR>` is
`<vector>`; FILE_ATTRIBUTE_NORMAL joins the shim with its Win32 value;
the CImm stub gains the no-argument Enable() two callers use.

Test fixes, all Windows assumptions in the tests themselves: the Assert
macro in PacketAssert.h and Assert1.h is one spelling on every platform
(the POSIX branch passed __PRETTY_FUNCTION__ and produced a different
message from the one test_packet_source_location.cpp pins); the
directory-listing test appends the platform's separator, since a
backslash is a filename character on POSIX; and TextBackendSDL.cpp's
font list gains the Linux locations of Noto CJK and DejaVu, without
which no font loads and the raster test cannot draw. That file was one
of the three CP949-encoded sources and is UTF-8 now, with its Korean
comment translated; MPlayer.h and MFakeCreature.h turned out to be
UTF-8 already except for one garbled comment tail each, repaired.

Test path: lib + test (packetwire, SpriteLib, gamemodel;
test_wire_layout.cpp, test_cspritepalbase.cpp).

Verified: Windows Debug and ASan trees 0 errors, 601 tests, 294,423
checks, 0 failed, seven ctests green. Linux, Ubuntu 24.04 in a
container: GCC 13 and Clang 18 build every library and both test
binaries with 0 errors, 601 tests, 294,422 checks, 0 failed, all seven
ctests pass (ratchets, arch_includes, format_arity, packet_indices,
wire_inventory_fresh included); the ASan+UBSan tree the same with 0
sanitizer reports. tests/golden is byte-identical to the server repo's
shared files and GCShopBought is 284 in both wire-layout.txt files.

Out of scope: the DarkEden executable does not link on Linux yet
(areas D and E), and the remaining wire-layout.txt differences against
the server are the pre-existing ones its task 1.4 tracks.

+++ Client/Packet/Types/ShopTypes.h
@@ -32,9 +32,12 @@ const int SHOP_RACK_INDEX_MAX = 5*4; // width*height = 20
 
 
 //////////////////////////////////////////////////////////////////////////////
-// 상점 런타임 버전 체크에 쓰이는 자료형
+// The shop's runtime version, checked against the server's. int32_t, as
+// the server spells it: it was `long`, and sizeof(long) is 8 on LP64, so a
+// Linux client sized GCShopBought four bytes larger than the server
+// (tests/wire-layout.txt caught it). The wire has always been four bytes.
 //////////////////////////////////////////////////////////////////////////////
-typedef long ShopVersion_t;
+typedef int32_t ShopVersion_t;
 const int szShopVersion = sizeof(ShopVersion_t);
